fix: confusing label
Earlier the label was 'Sweep delay', and it was changed to 'Transmit
delay'. Neither of them communicate the feature clearly, as it's not
obvious what the delay is.

The label is for the time the device should transmit or receive on
a particular channel. Which is better communicated by using the term
period, as that communicates the time duration of an event.

Further specifying that the time unit should be the period per channel,
clearly communicates to the user what the variable is used for.